Revert "PR-1d: extract CudaRenderCtx out of dialect-level LoweringCtx (#81)"
This reverts commit 904f89f2f3e02c7563c43831947c67cfda825113.
